pyproject.toml
[project]
name = "navigation-server"
package-mode = true
version = "2.6.1"
description = "Navigation Server Full Install"
authors = [{name="Laurent Carré", email="<laurent.carre@sterwen-technology.eu>"}]
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE"}

dependencies = [
    "pyserial>=3.5",
    "protobuf==5.27.2",
    "grpcio==1.66.2",
    "pyyaml>=6.0.1",
    "pyubx2 (>=1.2.50,<2.0.0)"
]

[project.optional-dependencies]
gpio = ["gpiod>=2.0"]
develop = ["grpcio-tools==1.66.2"]
can = ["python-can>=4.0"]
i2c = ["pylibi2c"]

[project.scripts]
navigation_server = "navigation_server:server_main"

[tool.poetry]

include = [
    {path = "navigation_definitions", format = ["sdist", "wheel"] },
    {path = "conf", format = ["sdist"] },
    {path = "system", format = ["sdist"] },
    "run_server", "run_server.py", "setrunenv", "run_script"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

